requirements.txt
nemo-toolkit[asr]
bark
gradio
scipy
numpy
PyPDF2 